requirement.txt
Appium-Python-Client==0.11
certifi==2017.4.17
chardet==3.0.4
custom-text-test-runner==0.1.5
enum34==1.1.6
idna==2.5
PyHamcrest==1.9.0
requests==2.18.1
requests-file==1.4.2
selenium==3.4.3
selenium-requests==1.3
selenium-unittest==0.2.3
six==1.10.0
table-printer==0.1.4
tldextract==2.1.0
urllib3==1.21.1
